scxtop: use smartstr with SSO for comms
We pass around a bunch of Comm strings in scxtop that comfortably fit in the
allocation of the string, but were each doing an allocation/deallocation and
being copied multiple times.

Use a crate `smartstr` to get a string with Small String Optimization. This
string fits 23 bytes in the string itself without a heap allocation.

Test plan:
- Took a perfetto trace, comms look good.