scx_lavd: Enable SCX_OPS_ENQ_LAST.
Since the scheduler focuses on latency and interactivity, let the scheduler
judiciously decide the task's time slice even when there is no other task to
run. This avoids the case that the sched_ext core chooses the default time
slice, which is pretty long (20 ms). Such a case is bad regarding latency
because such tasks should be kicked out by the preemption logic when more
urgent tasks come in.

Signed-off-by: Changwoo Min <changwoo@igalia.com>