scx_lavd: Extend the overflow set at the enqueue path by core preference.
When a task could not run on the currently active/overflow set,
such a situation was found at the dispatch path, and the overflow
set was extended at the dispatch path. Instead of waiting until the
dispatch, let's eagerly detect such a situation at the enqueue path
and extend the overflow set. This will reduce the latency of extending
the overflow set and the scheduling delay for the task.

Signed-off-by: Changwoo Min <changwoo@igalia.com>