scx_utils: Fix tracepoint_exists()
In tracepoint_exists(), unwrap_or() was used to get a result of either
tracefs_mount() or debugfs_mount(). However, in Rust, unwrap_or() is eager
so both tracefs_mount() and debugfs_mount() are evaluated.
Hence, if one of them does not exist, an error is raised.

So, we replace unwrap_or() with unwrap_or_else(), which conducts lazy
evaluation, meaning that if tracefs_mount() succeeds, debugfs_mount()
is never evaluated.

Also, added "/tracing" to the debugfs mount point since the debugfs mount
point is "/sys/kernel/debug" not "/sys/kernel/debug/tracing".

Signed-off-by: Changwoo Min <changwoo@igalia.com>