scx_layered: Introduce StickyDynamic growth algorithm
Introduce a new growth algorithm that remains sticky to LLCs and
allocates them in weighted proportion according to the size of a given
layer. Core orders decide the order in which CPUs are allocated or
freed, and how many CPUs can be allocated as well (if the layer is not
Grouped or Open). By refreshing the core order on every step, we can
ensure that the LLC and core assignments can be reflected according to
the dynamic nature of changing layer sizes.

The algorithm is fairly simple, we compute target LLCs in each step() as
a function of weighted number of CPUs. Then, in the first pass, all
layers that shrunk by an LLC in the window free up their LLCs and
contribute them back to the free LLC pool in the CpuPool. Next, we will
allocate LLCs to the layers that grew by an LLC. For extra cores less
than the number of cores in an LLC, we allocate them according to the
layer size from the remaining free LLCs. The computation of target LLCs
and extra cores always adds up to the total number of cores, thus, we
don't need to handle cases where extra cores are to be allocated but we
have no free LLCs to draw them from.

Right now, the algorithm has two levels (free/alloc LLCs, then do the
same for individual cores), but we can add an extra level for layers
that occupy cores at NUMA granularity, before stepping a level down into
extra LLCs (that don't add up to a NUMA node), and extra cores
thereafter.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>