scx_bpfland: Restore the correct value of pm_qos_resume_latency_us
When --idle-resume-us is used, we are changing the global system
settings of the idle CPU resume latency and when the scheduler is
stopped we are implicitly resetting the latency to 5000 us, assuming
that is the default.

However, the default can be different across architectures or it can be
explicitly set by the user.

Therefore, make sure to restore the same previous value of
pm_qos_resume_latency_us when the scheduler is stopped.

Fixes: abbc4c23 ("scx_bpfland: Introduce --idle-resume-us")
Signed-off-by: Andrea Righi <arighi@nvidia.com>