scx_bpfland: Normalize task weight
The task weight range is quite large: [1 .. 10000] (default 100) and it
can occasionally lead to stalls when tasks with low nice values run
alongside those with high nice values. This issue arises from the way
vruntime, and consequently task's deadline, is scaled.

For example, in Fedora with dnf-automatic enabled, the dnf process may
run with a niceness of 19. If there are tasks running with nice values
of -10 or -20, their relative vruntime scaling factor would be 1000 or
even 10000, which can easily increase the risk of triggering stalls.

To mitigate the influence of task weight on deadline computation,
normalize the weight into a smaller range: [1 .. 128] (default 64).

This normalization ensures that the relative vruntime scaling factor in
the previous example would be around 20 or 100 (at most), instead of
1000 and 10000.

This ensures a more balanced scheduling, influenced more by the task's
behavior rather than its priority, and it can also help to avoid
potential stall conditions in presence of large priority gaps.

Signed-off-by: Andrea Righi <arighi@nvidia.com>