scx_layered: Account used CPU time from ops.tick()
scx_layered was accounting used CPU time from ops.stopping() which has the
following problems:

- On newer kernels, ops.running/stopping() aren't called when a task keeps
  running on the same CPU. This can lead to large accumulation of
  unaccounted runtime.

- Even if CPU time is accounted every slice, with enough number of CPUs, the
  compound error can be substantial as typical slice durations are
  signficant fractions of the userspace scheduling interval (100ms by
  default).

As a work around, userspace was trying to compensate for the currently
running task but the implementation was racy and the result can go wildly
wrong on large machines. This led to wild fluctuations of CPU allocations on
large machines which likely forced most configs to limit the range with
explicit cpu_ranges config.

ops.tick() has been around for a while now. Let's account from there so that
the usage stats reliably stay uptodate.