scx_lavd: Directly dispatch a task when a system is idle enough.
At ops.enqueue() path, pick_proper_dsq() finds a proper compute domain (DSQ)
and a CPU where a task would run.

In case the chosen CPU is idle, and there is no remaining task on the compute
domain, it is safe to directly dispatch the task to the chosen CPU's local DSQ.
This allows us to bypass the unnecessary steps of placing the task to the
domain's DSQ and then moving it to the local DSQ again, thus reducing the
latency and keeping the CPU busy.

Moreover, there is no priority inversion in the domain since there is no
waiting task in the domain's DSQ.

Signed-off-by: Changwoo Min <changwoo@igalia.com>