scx_lavd: Relax the atomic visibility of sys_stat.
Initially, we managed two copies of sys_stat (i.e., like double buffering) to
guarantee the atomic visibility of sys_stat. However, over time, it turned out
that it was too much because there was only one writer to sys_stat, and the
write interval was tens of the msec scale. Moreover, even if the atomic
visibility is violated, the error is transient and the code is resilient to
such errors. So let's use only one copy of sys_stat for simplicity.

Signed-off-by: Changwoo Min <changwoo@igalia.com>