scx_lavd: Do not proactively reserve overflow CPUs.
Previously, one overflow CPU was allocated by default to avoid overutilizing
active CPUs proactively. However, it is not necessary in practice because
there are almost always affinitized tasks, causing overflow CPUs to be assigned.
So, drop proactive allocation of an overflow CPU and let's rely on the real
overflow CPUs.

Signed-off-by: Changwoo Min <changwoo@igalia.com>