scx_bpfland: Rework preemption behavior
Allowing tasks to interrupt others before their time slice expires is
often introducing more issues than benefits, specifically when the
system is overloaded we may trigger a storm of context switches, that
are never beneficial and can lead to massive system lags.

Therefore, rework preemption by tying it to the concept of direct
dispatch: tasks are always allowed to run for their entire assigned time
slice and the --no-preempt option effectively disables direct
dispatching.

In this context, preemption can be interpreted as the ability to
interrupt or bypass the regular scheduling flow.

Disabling direct dispatches (now controllable via --no-preempt) can help
to improve fairness at the cost of system responsiveness, which can be
an acceptable trade-off for certain server workloads.

Tested-by: Piotr Gorski <piotrgorski@cachyos.org>
Signed-off-by: Andrea Righi <arighi@nvidia.com>