chaos: add --ppid-required flag
A feature to only affect tasks in the process under test has been requested in
chaos. The implementation chosen here is filtering based on having a specific
parent somewhere in the process's lineage, with the option to include that parent
in the chaos or not. This new setting defaults to on when you provide a process
for the scheduler to manage or a PID to track, but can be disabled.

The implementation walks the `real_parent` path to root for the process, marking
it as `CHAOS_MATCH_HAS_PARENT` if it finds the target PPID or another process with
`CHAOS_MATCH_HAS_PARENT`. The implementation of `chaos_match` uses a bitset so
we can test more parameters in future.

Test plan:
- `-p` argument isn't implemented, this is tech debt for later!
- Added some logging: `scx_chaos -vv -- sh -c 'for i in {1..120}; do ( sleep 1 & ); sleep 1; done; wait; exit 0'` matches the shell and children but not chaos itself.
- `scx_chaos --ppid-targeting=false -vv -- sh -c 'sleep 120'` matches everything on the box.
- `scx_chaos` matches everything on the box.