scx_lavd: Use 1024 instead 1000 for fixed point arithmetic.
1000 has been used for fixed point arithmetic since it is easier to
understand for humans. Change to 1024 because 1) it is more efficient in
the machine and 2) it is compatible with the kernel conventions.

LAVD_SCALE and LAVD_SHIFT are defined and used throughout the code.
p2scale() and s2percent() are provided and used for conversion between
scale (1024) and percent (100).

There is no functional change.

Signed-off-by: Changwoo Min <changwoo@igalia.com>