rustscheds: add common SCX_TIMEOUT_MS env var to set stall watchdog limit
Currently the stall watchdog timer is either hardcoded into each scheduler's
struct_ops or left at the kernel default of 30s. This is inconvenient
particularly when performing testing as 30s is an incredibly long time.

Add an environment variable SCX_TIMEOUT_MS that dictates `timeout_ms` when set.
`scx_ops_open` is quite opinionated already, so we add it there and this variable
will propagate to all schedulers that use that. Will propose a similar change to
C-schedulers in the future.

Test plan:
- `cargo build --release && sudo SCX_TIMEOUT_MS=5000 target/release/scx_chaos --random-delay-frequency 0.1 --random-delay-min-us 5000000 --random-delay-max-us 5000000` - kicked for stalling with the environment variable, stays without.