scx_rustland_core: Proactively wakeup task's CPU in ops.enqueue()
If the scheduler is congested all the CPUs are already used, so there is
no practical benefit trying to wakeup idle CPUs.

Instead, it is more effective to proactively wakeup the CPU previously
used by the task. Even if the task needs to processed by the user-space
scheduler, having its CPU ready to consume the task can help to speedup
the entire BPF -> user-space -> BPF scheduling pipeline.

Signed-off-by: Andrea Righi <arighi@nvidia.com>