scx_rustland_core: Maximize CPU utilization
To avoid re-triggering the user-space scheduler in a busy loop when
there are pending tasks, we use an atomic flag (usersched_needed) and a
test-and-clear approach, allowing only the first CPU that checks the
flag to trigger the user-space scheduler when needed.

However, this synchronization introduces bubbles in the scheduling
pipeline, leading to sub-optimal CPU utilization, especially under full
system saturation.

For example, a parallel kernel build `make -j$(nproc)` on a 8-CPUs
system shows the following (using scx_rustland):

  0[||||||||||||||||||||||||98.8%]   4[||||||||||||||||||||||||98.2%]
  1[||||||||||||||||||||||||98.8%]   5[||||||||||||||||||||||||98.2%]
  2[||||||||||||||||||||||||98.8%]   6[||||||||||||||||||||||||98.2%]
  3[||||||||||||||||||||||||98.2%]   7[||||||||||||||||||||||||98.8%]

To eliminate these bubbles, increase the granularity of the check for
pending tasks by adding an additional check in ops.dispatch(), right
before the test-and-clear operation.

This allows CPUs to detect pending tasks earlier, while still ensuring
that only one CPU actually fires the user-space scheduler, avoiding
multiple unnecessary dispatches of the user-space scheduler.

With this change applied, the CPU utilization during the parallel kernel
build looks like the following:

  0[|||||||||||||||||||||||100.0%]   4[|||||||||||||||||||||||100.0%]
  1[|||||||||||||||||||||||100.0%]   5[|||||||||||||||||||||||100.0%]
  2[|||||||||||||||||||||||100.0%]   6[|||||||||||||||||||||||100.0%]
  3[|||||||||||||||||||||||100.0%]   7[|||||||||||||||||||||||100.0%]

In both cases the CPU utilization of the user-space scheduler remains
the same.

Signed-off-by: Andrea Righi <arighi@nvidia.com>