ci: squash layered matrix
Deduplicate the code in integration-tests.yml more by removing
layered-matrix. Add a code generator for the test matrix and use that to
decide which tests to run. This allows having a layered matrix within
the same list. Add `flags` as a parameter so we can feed extra flags to
any scheduler that needs it in the future.

This also removes one needed runner, as one of the layered matrix runs
before duplicated the integration test for layered.

Test plan:
- CI