scx_layered: Create pinned task local storage map
Add support for creating and pinning a 'scx_layered_task_hint_map' in
bpffs so that tasks on the system can open it and update their values
to add task hints. Mark the map as 666 so that all tasks are allowed
to open it and write to it by default.

As a follow up, we would want to enforce access control such that a
bpf_map_update_elem on this map can only happen from a task when the
pidfd for the key is the same as its thread ID. This needs to be done
using BPF LSM or something similar.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>