scx_tickless: Avoid distributing tasks on scheduling CPUs
When assigning tasks to the available CPUs, treat the scheduling CPUs as
a last resort to ensure they're kept free whenever possible. In this way
they can be used primarily for scheduling purposes, giving tasks a
better chance to run uninterrupted.

Moreover, make sure to mark CPUs that are used or skipped as busy to
prevent repeated attempts to dispatch tasks to the same CPU

Fixes: aed4f584 ("scx_tickless: Improve scheduling timer")
Signed-off-by: Andrea Righi <arighi@nvidia.com>