scx_utils: cast ioctl opcodes to libc::Ioctl
The opcode type that ioctl() accepts can differ between platforms,
namely between glibc where it accepts an unsigned 32-bit int, and musl
where it accepts a signed 32-bit int. Either way, the underlying value
of the opcode is a 32-bit integer. Currently, bindgen is storing the
enum values defined in perf_bindings.h as u32, which is fine for glibc,
but not musl (which wants an i32). This commit casts the opcodes to
libc::Ioctl before passing them to fix this.