scx_utils: Choose a CPU capacity source more judiciously.
A case was reported that /sys/devices/system/cpu/cpuX/acpi_cppc/highest_perf
have the same value for all CPUs on an Intel hybrid processor (i7-13650hx).
This should be a bug on the driver's side.

To circumvent such a problem, make more efforts to find a source that can tell
the capacity differences among the cores. If there is no such source, resort
to the max frequency of CPUs to estimate CPU capacity.

This should solve the following issue:
  https://github.com/sched-ext/scx/issues/1893

Signed-off-by: Changwoo Min <changwoo@igalia.com>