scx_lavd: Count runtime only once in lat_cri calculation.
Previously, the average runtime of a task is counted double in calculating
the task's latency criticality. However, double-counting task runtime does
not show clear benefits, so let's simplify the equation.

Also, limit the accumulation of runtime up to the max time slice, so
preferring a recently woken-up task is limited for one max time slice.

Signed-off-by: Changwoo Min <changwoo@igalia.com>