scx_rustland_core: Introduce a separate DSQ for the user-space scheduler
The user-space scheduler is always dispatched after all the other queued
tasks have been dispatched.

This ensures to work in bursts: tasks are queued, then the user-space
scheduler runs and dispatches them. Once all these tasks exhaust their
time slices, the scheduler is invoked again, repeating the cycle.

This is implemented dispatching the user-space scheduler itself to the
SHARED_DSQ, assigning the biggest vtime possible, so that it always runs
last. This is not straighforward to understand and it potentially
introduces unnecessary overhead to manage the artificial vtime.

Therefore, introduce a separate DSQ to dispatch the user-space
scheduler, used in FIFO mode, that will be always consumed after all the
other DSQs.

Basically the behavior is the same, but this makes the whole BPF /
user-space workflow easier to understand and it also reduces a bit of
overhead.

Tested-by: Mario Roy <marioeroy@gmail.com>
Signed-off-by: Andrea Righi <arighi@nvidia.com>