chaos: refactor trait selection
Chaos trait selection is currently handwritten for each trait. Refactor this to
have a uniform selection process based on the `chaos_trait` value. This makes adding
types easier in the future.

This also implies that CHAOS_TRAIT_NONE is the most unlikely, and this is now
explicitly selected as the first random value. Also uses an unrolled loop in the
hope that the `const volatile` can be JIT'ted nicely, but this isn't overly
important.

Test plan:
- CI
- Manually, tested a few and read the debug output which looks good.