scx_bpfland: Ignore sync wakeup with short-running wakers
During a sync wakeup the waker is going to release the CPU immediately
after the wakeup. which often reflects a producer/consumer interaction.

In such cases, the waker may have produced data that the wakee will
consume, making it potentially beneficial to migrate the wakee to the
waker's CPU. This speculation assumes that relevant cache-hot data
resides on the waker's CPU, potentially more so than on the wakee's
previous CPU.

However, if the waker has only been running briefly since its last
sleep, it’s likely that little useful data has been produced in its
CPU's cache. In this scenario, the wakeup is probably more of a pure
synchronization event (signaling without significant data transfer),
making migration less beneficial and the wakee might be better off
continuing the execution on its previous CPU.

To make this distinction, use the waker's exec_runtime (the time it has
run since its last sleep) as a heuristic to determine whether the sync
wakeup presents a good opportunity to migrate the wakee or not.

Signed-off-by: Andrea Righi <arighi@nvidia.com>