ci: build all kernels with nix before starting jobs
Currently we build a single CI kernel before starting a job. This means
that we can't reliably use alternative kernels in a single job, for
example if a scheduler requires a specific kernel, or if we want to
require jobs testing against bpf/bpf-next when that kernel gets updated.

Refactor the build-kernels job into a reusable workflow (again I
think...) so it's simpler to use. Replace the previous build-kernel
action's invocations with the explicit `nix build` commands in each
job, which allows us more flexibility when we need multiple kernels at
once in the future.

Drop the kernel build from update-kernels.yml as it's implied by the PR
and it's better for the PR to fail to merge while being visible.

Test plan:
- CI