scx_utils: Avoid divide-by-zero when "PSS Support" is disabled.
When "PSS Support" is disabled in the BIOS, there are no CPU frequency and
capacity-related files under sysfs, as the OS does not have control over the
processor's performance and power. In this case, the max_rcap of the previous
code will be zero, causing a divide-by-zero error. To avoid this, let's choose
the default value (1024) for max_rcap, etc. when no capacity-related
information is available.

Signed-off-by: Changwoo Min <changwoo@igalia.com>