scx_layered: Replenish slice for protected layers
When there's excessive yields from a task that is not sharing its CPU
with others, and yield_ignore is not 1, we enter the dispatch path and
fail the keep_running() check, but the expectation is that, after
checking our DSQs, we'll end up replenishing the slice, because we're
the only task running on this CPU. However, we short circuit layer DSQ
draining for protected layers, as we don't want to pull tasks from other
layers onto our CPUs, and only check and consume our own DSQs.

Returning early causes the process to be enqueued again instead of being
reenqueued into the local DSQ by default, since it's time slice has been
exhausted. We must replenish it before return to restore the correct
behavior.

Given we would have checked lo_fb again if this worked properly, place
the jump label above the tried_lo_fb check and test it, and then fall
through to the point where we replenish the slice for the previous task.

Fixes: c37b64c5bf90 ("scx_layered: introduce protected layers")
Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>