scx: add a PoC unit test with supporting code for scx_p2dq
The vast majority of the scheduler logic for scx schedulers exists in
bpf code where there's no easy way to unit test.

To address this we will have to stub out and create user space
implementations of various helpers.

To get the ball rolling I've added a basic test for a function in
scx_p2dq that involves no bpf helpers.  This gives us the bare minimum
we need in order to compile bpf code as userland C code, and then run it
as a test.

You can test this with

meson test -C build

Further work will be to flesh out the helpers and add a lot more
functionality.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>