scx_flash: Update nvcsw statistics only on task sleep
ops.quiescent() is typically called when a task goes to sleep, but it
may also be invoked when a task is dequeued due to changes in its
properties, such as CPU affinity updates or cgroup migrations.

Ensure that the voluntary context switch rate is updated only in
response to genuine sleep events.

Signed-off-by: Andrea Righi <arighi@nvidia.com>