scx_lavd: Use perf_cpu_order for the BPF scheduler.
Use perf_cpu_order generated by either leveraging the energy model, if
available or by using the system's CPU topology.

When there are more than two performance levels, we ignore the middle levels
and use only the first as the power save mode and the last as the performance
mode for now. This limitation will be removed in the following commits, along
with proper changes to the BPF code.

No functional & behavioral changes are expected when the energy model is not
available.

Signed-off-by: Changwoo Min <changwoo@igalia.com>