scx_lavd: Use the PCO table when calculating the number of active CPUs.
When the energy model is available, use the upper bounds of the lowest
performance state and the lower bounds of the highest performance state
as the thresholds for the power save and performance modes, respectively.
Use the PCO table so that we can also utilize the middle states in
the PCO table.

When switching to a different performance mode (e.g., power save -> balanced),
use the required compute capacity, which is compatible with the PCO table,
instead of relying on CPU utilization based on wall-clock time. Clean up the
calculation of avg_sc_util to use more time_delta() and remove redundant reads
of volatile variables, etc.

The old, unnecessary code is removed as now the code compiles successfully.
Now the scheduler uses the first PCO entry for powersave mode and the last
PCO entry for performance mode. It does not use the states in the middle yet.

Signed-off-by: Changwoo Min <changwoo@igalia.com>