scx_flash: Remove the tickless timer
Replace the periodic timer used to check for contended CPUs with a
direct check in ops.enqueue(): when enqueuing a task, if its previously
used CPU is running a tickless task (a task with an infinite time
slice), assign it a regular time slice. This resumes the tick on that
CPU, effectively re-enabling preemption for contended CPUs.

With this, the global tickless timer is no longer needed. Consequently,
the --tickless_ns option can be simplified to a --tickless boolean flag.

Suggested-by: Changwoo Min <changwoo@igalia.com>
Signed-off-by: Andrea Righi <arighi@nvidia.com>