scx_lavd: Handling the CPU utilization spikes in autopilot mode.
When a CPU is almost fully utilized (say 90%) during a period, we may
underestimate the utilization. For example, when the measured CPU
utilization is 100%, there is a possibility that the actual utilization
is actually higher, such as 130%.

To handle such utilization spike cases, we give a 50% premium for the
scaled CPU time where the CPU is almost fully utilized. This overestimates
the scaled CPU utilization and the required compute capacity. Then, this
finally allocates more active CPUs. The over-allocated CPUs become the
breathing room.

Signed-off-by: Changwoo Min <changwoo@igalia.com>