scx_flash: Introduce --sticky-cpu
Introduce the new option --sticky-cpu to allow reusing the same CPU in
ops.select_cpu() even when the CPU is outside of the primary scheduling
domain.

This can help reduce task migrations (which can be benficial for server
workloads) at the cost of introducing less consistent performance in
presence of systems with hybrid cores.

Tested-by: Piotr Gorski <piotrgorski@cachyos.org>
Signed-off-by: Andrea Righi <arighi@nvidia.com>