ci: add required-checks job
GitHub branch protection rules aren't defined in the code and changes to
them have to be made by administrators instead of the usual code review
process. This is inconvenient for two reasons: external contributors
can't help with many CI changes, and any changes to them require every
PR to be rebased before it'll pass the CI.

Add a `required-checks` job to replace the `rust-tests` and
`integration-test/all-success` requirements currently in the branch rule.
`integration-test/all-success` already partially covers this for
adding/removing integration tests, but we don't have an equivalent for
caching-build. Adding this job and making it the requirement makes
splitting `rust-tests` for better parallelism less annoying, for
example. Will enable this check in the CI after this has bedded in for a
few days so most contributions will already have it.

Test plan:
- CI